requirements.txt
streamlit
PyPDF2
python-docx
cerebras-cloud-sdk

